moving test outside of package using different flag
