+++ pkg/tests/apis/provisioning/git/missing_metadata_incremental_test.go
@@ -1,0 +1,56 @@
+package git
+
+import (
+	"strings"
+	"testing"
+
+	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/pkg/tests/apis/provisioning/common"
+	gitcommon "github.com/grafana/grafana/pkg/tests/apis/provisioning/git/common"
+	"github.com/grafana/grafana/pkg/util/testutil"
+	"github.com/stretchr/testify/require"
+	"k8s.io/apimachinery/pkg/runtime"
+)
+
+func TestIntegrationProvisioning_IncrementalSync_MissingFolderMetadata_FlagDisabled(t *testing.T) {
+	testutil.SkipIntegrationTestInShortMode(t)
+
+	helper := gitcommon.RunGrafanaWithGitServer(t) // no withProvisioningFolderMetadata
+
+	const repoName = "incr-missing-meta-disabled"
+
+	_, local := helper.CreateGitRepo(t, repoName, map[string][]byte{
+		"dashboard.json": gitcommon.DashboardJSON("disabled-dash", "Root Dashboard", 1),
+	})
+
+	// Full sync.
+	common.SyncAndWaitWithSuccess(t, helper, repoName)
+
+	// Add a dashboard inside a folder with no _folder.json.
+	require.NoError(t, local.CreateFile("myfolder/dashboard2.json", string(gitcommon.DashboardJSON("disabled-folder-dash", "Folder Dashboard", 1))))
+	_, err := local.Git("add", ".")
+	require.NoError(t, err)
+	_, err = local.Git("commit", "-m", "add dashboard in folder without metadata")
+	require.NoError(t, err)
+	_, err = local.Git("push")
+	require.NoError(t, err)
+
+	// Trigger incremental sync.
+	job := helper.TriggerJobAndWaitForComplete(t, repoName, provisioning.JobSpec{
+		Action: provisioning.JobActionPull,
+		Pull:   &provisioning.SyncJobOptions{Incremental: true},
+	})
+	jobObj := &provisioning.Job{}
+	require.NoError(t, runtime.DefaultUnstructuredConverter.FromUnstructured(job.Object, jobObj))
+
+	require.Empty(t, jobObj.Status.Errors,
+		"incremental sync with flag disabled should produce no errors")
+	require.Equal(t, provisioning.JobStateSuccess, jobObj.Status.State,
+		"incremental sync should succeed without warnings when flag is disabled")
+
+	// Ensure no warning about missing folder metadata.
+	for _, w := range jobObj.Status.Warnings {
+		require.False(t, strings.Contains(w, "missing folder metadata"),
+			"should not warn about missing folder metadata when flag is disabled, got: %s", w)
+	}
+}
